+++ packages/studio-query/src/app/content/index.tsx
@@ -1,6 +1,8 @@
-import React, { useState, memo } from 'react';
-import Statement from '../../statement';
-import Header from './header';
+import React, { useState, memo, lazy, Suspense } from 'react';
+// import Statement from '../../statement';
+const Statement = lazy(() => import('../../statement'));
+// import Header from './header';
+const Header = lazy(() => import('./header'));
 import { Segmented, theme } from 'antd';
 import { useContext } from '../context';
 import type { IStatement, IStudioQueryProps } from '../context';
@@ -86,7 +88,10 @@ const Content: React.FunctionComponent<IContentProps> = props => {
           borderBottom: `1px solid ${token.colorBorder}`,
         }}
       >
-        <Header />
+        <Suspense>
+          <Header />
+        </Suspense>
+
         {mode === 'tabs' && queryOptions.length !== 0 && (
           <div style={{ padding: '8px 0px' }}>
             <Segmented options={queryOptions} onChange={handleChangeTab} value={activeId} />
@@ -107,22 +112,24 @@ const Content: React.FunctionComponent<IContentProps> = props => {
                 flex: 1,
               }}
             >
-              <Statement
-                language={language}
-                enableImmediateQuery={enableImmediateQuery}
-                mode={mode}
-                active={id === activeId}
-                saved={savedIds.indexOf(id) !== -1}
-                id={id}
-                timestamp={timestamp}
-                graphName={graphName}
-                schemaData={schemaData}
-                script={script}
-                onQuery={queryGraphData}
-                onCancel={handleCancelQuery}
-                onClose={onClose}
-                onSave={onSave}
-              />
+              <Suspense>
+                <Statement
+                  language={language}
+                  enableImmediateQuery={enableImmediateQuery}
+                  mode={mode}
+                  active={id === activeId}
+                  saved={savedIds.indexOf(id) !== -1}
+                  id={id}
+                  timestamp={timestamp}
+                  graphName={graphName}
+                  schemaData={schemaData}
+                  script={script}
+                  onQuery={queryGraphData}
+                  onCancel={handleCancelQuery}
+                  onClose={onClose}
+                  onSave={onSave}
+                />
+              </Suspense>
             </div>
           );
         })}
