+++ packages/studio-query/src/app/content/loading.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+export default function Loading() {
+  return (
+    <p>
+      <i>Loading...</i>
+    </p>
+  );
+}
