+++ packages/studio-website/package.json
@@ -11,6 +11,7 @@
     "build": "umi build --engineType=interactive",
     "build:interactive": "umi build  --engineType=interactive",
     "build:groot": "umi build --engineType=groot",
+    "build:analysis": "ANALYZE=1 umi build  --engineType=interactive",
     "lint": "umi lint"
   },
   "dependencies": {
